refactor: import sleep directly; dedup randomIndex
Drop the clock.js re-export — consumers import src/sleep.js directly
(global-setup already did). Extract the byte-identical _randomIndex
bodies in Bottleneck and Job into src/random-index.js. No utils junk
drawer: one concept per module stays the house style.
